analytics_service/requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
pymongo==4.6.1
schedule==1.2.1
python-dotenv==1.0.0